Use interface for singleflight group
